refactor: rename AuthContext.jsx to authStore.js for consistent architecture (#221)
- Move src/contexts/AuthContext.jsx → src/stores/authStore.js
- Update all import statements in useAuthFlow, usePasswordRotation, useDataManagement
- Remove empty src/contexts/ directory
- Maintain consistent Zustand store naming pattern (budgetStore.js, authStore.js)
- Preserve AuthProvider component for auto-logout functionality
- All exports and functionality remain identical

✅ All imports resolve correctly
✅ Linting warnings reduced from 20 to 19
✅ Consistent /stores/ directory structure maintained
