+++ src/hooks/useAuthFlow.js
@@ -1,5 +1,5 @@
 import { useCallback } from "react";
-import { useAuth } from "../contexts/AuthContext";
+import { useAuth } from "../stores/authStore";
 import logger from "../utils/logger";
 
 /**
